pyproject.toml
[project]
name = "odoo-order-poller"
version = "0.1.0"
description = "Microservicio que hace polling a instancias Odoo SaaS y envía órdenes como webhooks"
requires-python = ">=3.12,<4.0"
dependencies = [
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
odoo-poller = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
